+++ src/main/java/com/mraof/minestuck/player/IdentifierHandler.java
@@ -1,6 +1,8 @@
 package com.mraof.minestuck.player;
 
+import com.mojang.serialization.Codec;
 import com.mojang.serialization.DataResult;
+import com.mojang.serialization.codecs.RecordCodecBuilder;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.Tag;
 import net.minecraft.server.MinecraftServer;
@@ -17,6 +19,7 @@ import java.util.*;
 /**
  * Used to encode/decode player usernames, to handle uses with LAN.
  * This file is to now only be used serverside.
+ *
  * @author kirderf1
  */
 public class IdentifierHandler
@@ -39,7 +42,7 @@ public class IdentifierHandler
 		return identifier;
 	}
 	
-	@Deprecated	//Prefer the DataResult functions
+	@Deprecated    //Prefer the DataResult functions
 	@Nonnull
 	public static PlayerIdentifier loadOrThrow(CompoundTag tag, String key) throws RuntimeException
 	{
@@ -159,11 +162,17 @@ public class IdentifierHandler
 		fakePlayerIndex = 0;
 	}
 	
-	private static class UUIDIdentifier extends PlayerIdentifier
+	public static class UUIDIdentifier extends PlayerIdentifier
 	{
+		public static final Codec<UUIDIdentifier> CODEC = RecordCodecBuilder.create(instance -> instance.group(
+				Codec.INT.fieldOf("id").forGetter(UUIDIdentifier::getId),
+				Codec.STRING.xmap(UUID::fromString, UUID::toString).fieldOf("uuid").forGetter(UUIDIdentifier::getUUID)
+		).apply(instance, UUIDIdentifier::new));
+		
 		private final UUID uuid;
 		
-		private UUIDIdentifier(int id, UUID uuid)
+		//TODO make this constructor private again. Exposed for testing
+		public UUIDIdentifier(int id, UUID uuid)
 		{
 			super(id);
 			this.uuid = uuid;
@@ -178,7 +187,12 @@ public class IdentifierHandler
 		@Override
 		public String getUsername()
 		{
-			return UsernameCache.containsUUID(uuid) ? UsernameCache.getLastKnownUsername(uuid) : "Unknown ("+getId()+")";
+			return UsernameCache.containsUUID(uuid) ? UsernameCache.getLastKnownUsername(uuid) : "Unknown (" + getId() + ")";
+		}
+		
+		public UUID getUUID()
+		{
+			return uuid;
 		}
 		
 		@Override
@@ -207,7 +221,7 @@ public class IdentifierHandler
 		@Override
 		public String toString()
 		{
-			return "Identifier:"+getUsername();
+			return "Identifier:" + getUsername();
 		}
 		
 		@Override
@@ -245,7 +259,7 @@ public class IdentifierHandler
 		@Override
 		public String getUsername()
 		{
-			return "Fake player "+count;
+			return "Fake player " + count;
 		}
 		
 		@Override
@@ -257,21 +271,21 @@ public class IdentifierHandler
 		@Override
 		public String getCommandString()
 		{
-			return "fake"+count;
+			return "fake" + count;
 		}
 		
 		@Override
 		public CompoundTag saveToNBT(CompoundTag nbt, String key)
 		{
 			nbt.putString(key, "fake");
-			nbt.putInt(key+"_count", count);
+			nbt.putInt(key + "_count", count);
 			return nbt;
 		}
 		
 		@Override
 		public String toString()
 		{
-			return "Identifier:fake_"+count;
+			return "Identifier:fake_" + count;
 		}
 		
 		@Override