+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Block Teleporter
 - Built in compatability for Better Combat mod
+- GristTorrent Computer App
 
 ### Changed
 - Structure Block Registry Processor can now be used for data generated structures through a processor_list